Convert resources.py to typesafe imple
